Cli rewrite achieved. - core/install & core/uninstall removed.
